requirement.txt
accelerate==0.30.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.2.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
fonttools==4.53.1
frozenlist==1.7.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.13.4
idna==3.7
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.11
langsmith==0.4.29
lz4==4.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.108.1
openpyxl==3.1.5
orjson==3.11.3
packaging==24.1
pandas==2.3.0
pillow==10.4.0
pinecone==7.0.1
pinecone-plugin-interface==0.0.7
propcache==0.3.2
protobuf==5.29.4
protoc-gen-openapiv2==0.0.1
psutil==7.0.0
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.14.1
sentencepiece==0.2.0
simsimd==6.5.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.5.0
tiktoken==0.11.0
tokenizers==0.13.3
torchvision==0.22.0
tqdm==4.67.1
transformers==4.29.2
triton==3.3.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.20.1
zstandard==0.25.0
